+++ app/helpers/contacts_helper.rb
@@ -1,0 +1,2 @@
+module ContactsHelper
+end
