+++ app/models/contact.rb
@@ -1,3 +1,3 @@
-class Status < ApplicationRecord
+class Contact < ApplicationRecord
   validates :name, :message, presence: true
 end
