+++ app/views/status/index.html.erb
@@ -47,10 +47,5 @@
         </div>
       </div>
     </div>
-    <div class="status__box_lower">
-      <div class="status__box_lower_comment">
-        <%= form_with %>
-      </div>
-    </div>
   </div>
 </div>
