+++ config/routes.rb
@@ -1,5 +1,5 @@
 Rails.application.routes.draw do
   root to: "status#index"
-  resources :status, only: [:index, :create]
-
+  resources :status, only: [:index]
+  resources :contact, only: [:new, :create]
 end
