+++ test/controllers/contacts_controller_test.rb
@@ -1,6 +1,6 @@
 require 'test_helper'
 
-class StatusTest < ActiveSupport::TestCase
+class ContactsControllerTest < ActionDispatch::IntegrationTest
   # test "the truth" do
   #   assert true
   # end
