+++ test/models/contact_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class ContactTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
